refactor(auth): update exception handling
